First parallel implementation, different approaches (iteration_step variations) for BBDT3D

+++ src/GRAPHGEN/base_ruleset.h
@@ -95,7 +95,7 @@ public:
 
     virtual rule_set GenerateRuleSet() = 0;
 
-	virtual action_bitset GetActionFromRuleIndex(const rule_set& rs, uint rule_index) const = 0;
+	virtual action_bitset GetActionFromRuleIndex(const rule_set& rs, uint64_t rule_index) const = 0;
 
 };
