+++ src/Labeling/grana_ruleset.h
@@ -63,7 +63,7 @@ public:
     }
 
 
-	action_bitset GetActionFromRuleIndex(const rule_set& rs, uint rule_index) const override {
+	action_bitset GetActionFromRuleIndex(const rule_set& rs, uint64_t rule_index) const override {
 		rule_wrapper r(rs, rule_index);
 
 		bool X = r["o"] || r["p"] || r["s"] || r["t"];
